slightly smarter cleanup logic in e2e test

+++ test/test_e2e.py
@@ -24,10 +24,17 @@ class End2EndTestSuite(unittest.TestCase):
         headers = get_signed_headers(EMAIL)
 
         # cleanup resources
-        response = requests.delete(f"{TARGET_URL}/api/user_project/{ORG}/{PUBLIC_PROJECT_NAME}/data/{resource_type}", headers=headers)
-        self.assertTrue(response.status_code == 200 or response.status_code == 404)
-        response = requests.delete(f"{TARGET_URL}/api/user_project/{ORG}/{PUBLIC_PROJECT_NAME}/data/{resource_type}/generator", headers=headers)
-        self.assertTrue(response.status_code == 200 or response.status_code == 404)
+        response = requests.get(f"{TARGET_URL}/api/user_project/{ORG}/{PUBLIC_PROJECT_NAME}", headers=headers)
+        if response.status_code == 200:
+            response = requests.get(f"{TARGET_URL}/api/user_project/{ORG}/{PUBLIC_PROJECT_NAME}/data/{resource_type}", headers=headers)
+            if response.status_code == 200:
+                response = requests.delete(f"{TARGET_URL}/api/user_project/{ORG}/{PUBLIC_PROJECT_NAME}/data/{resource_type}", headers=headers)
+                self.assertTrue(response.status_code == 200 or response.status_code == 404)
+
+            response = requests.get(f"{TARGET_URL}/api/user_project/{ORG}/{PUBLIC_PROJECT_NAME}/data/{resource_type}/generator", headers=headers)
+            if response.status_code == 200:
+                response = requests.delete(f"{TARGET_URL}/api/user_project/{ORG}/{PUBLIC_PROJECT_NAME}/data/{resource_type}/generator", headers=headers)
+                self.assertTrue(response.status_code == 200 or response.status_code == 404)
 
         # response = requests.delete(f"{self.BASE_URL}/api/user_project/org123/project456/data_references", headers=headers)
         # self.assertEqual(response.status_code, 200)
